Move changeset bbox test to integration tests
